requirements.txt
backcall==0.2.0
clip @ git+https://github.com/openai/CLIP.git@573315e83f07b53a61ff5098757e8fc885f1703e
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
fonttools==4.28.2
ftfy==6.0.3
ipykernel==6.5.0
ipython==7.29.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pyzmq==22.3.0
regex==2021.11.10
scikit-learn==1.0.1
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.0
wcwidth==0.2.5
